Add Docker support for building and testing
The project now supports building and running inside a Docker container. This includes details on how to build the Docker image with verbose build output, run the container, and retrieve test results from the Docker container. This change also adds a Dockerfile and .dockerignore file to the project. Updates were also made to the Authlete.Tests project file.